qca-ssdk: Allow to update the existed MAPT decapsulation entry
The MAPT decapsulation entry can be repeatedly added with the same
source IPv6 address but different tunnel interface ID, as the tunnel
entry is configured in PPE when the tunnel interface is created, but
it can be deactive where the entry is not deleted, but the tunnel
interface ID will be different when the tunnel interface is active
again. To cover this case, when the MAPT entry is added with the
same IPv6 address as it is already configured in PPE, the same MAPT
entry index will be used to be added with this new MAPT entry.

Change-Id: I4b93ae00424ae6f6eea1cbdd8aaa371a225c92be
Signed-off-by: Luo Jie <quic_luoj@quicinc.com>
(cherry picked from commit c182b071e0fb42e658f6971e275d966d6e7fd0e5)